packages/dashboard/requirements.txt
streamlit>=1.44,<2
psycopg[binary]>=3.2,<4
pandas>=2.2,<3
python-dotenv>=1.0,<2
